pyproject.toml
[project]
name = "bulma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "google-generativeai>=0.8.4",
    "numpy>=2.2.3",
    "openai>=1.64.0",
    "pandas>=2.2.3",
    "pydantic-settings>=2.8.0",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
]
